+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.14.0
+
+* Reversing `structures/bk-tree#search` arguments.
+
 ## 0.13.0
 
 * Adding the `clustering/key-collision` namespace.